[refactor] Make collection wrappers take their marshal object.
This commit makes collections accept their marshal object, instead of
referencing a singleton. This will make it easier to make marshals
have a defined scope.
